Moved engine laser logic into Laser.ts
